Deleted personal info and added Documentation
