+++ nodejs/package.json
@@ -10,6 +10,6 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "axios": "^0.26.0"
+    "axios": "^0.26.1"
   }
 }
