+++ python/src/getDevices.py
